e_softhub/requirements.txt
asgiref
certifi
cffi
charset-normalizer
crispy-bootstrap5
cryptography
defusedxml
Django
django-allauth
django-crispy-forms
django-environ
django-filter
django-htmx
django-tinymce
djangorestframework
idna
Markdown
oauthlib
pillow
psycopg2
pycparser
PyJWT
python3-openid
pytz
requests
requests-oauthlib
six
sqlparse
tzdata
urllib3